- add feature to send emails  - change layers_views file to folder
