+++ layers/urls.py
@@ -1,14 +1,16 @@
 from django.urls import include, path, re_path
 from rest_framework.routers import DefaultRouter
-from layers.views import (
-    ListCreatePolygonLayer, RetrieveUpdateDestroyPolygonLayer,
-    RetrieveCreateUpdateUserDetail, FieldIndicatorsViewSet
-)
+from layers.views.contact_views import contact_us
+from layers.views.polygon_views import ListCreatePolygonLayer
+from layers.views.indicator_views import FieldIndicatorsViewSet
+from layers.views.polygon_views import RetrieveUpdateDestroyPolygonLayer
+from layers.views.user_detail_views import RetrieveCreateUpdateUserDetail
 
 router = DefaultRouter()
 router.register('fieldindicators', FieldIndicatorsViewSet, basename="fieldindicators")
 
 urlpatterns = [
+    path('contact/', contact_us),
     path('listcreatepolygonlayer/', ListCreatePolygonLayer.as_view()),
     path(
         'getupdatedeletelayer/<str:field_id>/',
