+++ requirements.txt
@@ -16,6 +16,7 @@ docopt==0.6.2
 idna==2.10
 isort==4.3.21
 lazy-object-proxy==1.4.3
+mailjet-rest==1.3.4
 Markdown==3.2.2
 mccabe==0.6.1
 numpy==1.19.1
