+++ sh_backend/urls.py
@@ -20,5 +20,4 @@ from django.urls import path
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('layers/', include('layers.urls')),
-    url(r'^api-auth/', include('rest_framework.urls'))
 ]
